requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
aiohttp==3.10.5
SQLAlchemy[asyncio]==2.0.31
asyncpg==0.29.0
motor==3.3.2
pydantic==2.8.2
pydantic-settings==2.10.1
psycopg2-binary
pillow
pymongo==4.6.0
python-multipart
# tests purpose only
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.24.0
pytest-asyncio>=0.21.0
